fix: move test to standalone module
